refactor: move createWizard to helpers
